# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AMHARC Match Capture — Local Agent API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /cameras:
    get:
      operationId: getCameras
      tags: [cameras]
      summary: List all configured cameras
      responses:
        "200":
          description: Camera list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Camera"
    post:
      operationId: createCamera
      tags: [cameras]
      summary: Add a new camera
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CameraInput"
      responses:
        "201":
          description: Camera created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Camera"
  /cameras/{cameraId}:
    get:
      operationId: getCamera
      tags: [cameras]
      summary: Get camera details
      parameters:
        - name: cameraId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Camera
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Camera"
        "404":
          description: Not found
    put:
      operationId: updateCamera
      tags: [cameras]
      summary: Update camera configuration
      parameters:
        - name: cameraId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CameraInput"
      responses:
        "200":
          description: Camera updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Camera"
    delete:
      operationId: deleteCamera
      tags: [cameras]
      summary: Remove a camera
      parameters:
        - name: cameraId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /cameras/{cameraId}/connect:
    post:
      operationId: connectCamera
      tags: [cameras]
      summary: Connect to camera
      parameters:
        - name: cameraId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Connection result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CameraConnectionResult"
  /cameras/{cameraId}/disconnect:
    post:
      operationId: disconnectCamera
      tags: [cameras]
      summary: Disconnect from camera
      parameters:
        - name: cameraId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Disconnected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperationResult"
  /cameras/{cameraId}/test:
    post:
      operationId: testCamera
      tags: [cameras]
      summary: Test camera connection
      parameters:
        - name: cameraId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Test result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CameraTestResult"
  /cameras/{cameraId}/ptz:
    post:
      operationId: sendPtzCommand
      tags: [cameras]
      summary: Send PTZ command to camera
      parameters:
        - name: cameraId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PtzCommand"
      responses:
        "200":
          description: PTZ command sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperationResult"
  /cameras/{cameraId}/presets:
    get:
      operationId: getCameraPresets
      tags: [cameras]
      summary: List PTZ presets for camera
      parameters:
        - name: cameraId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Presets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PtzPreset"
    post:
      operationId: saveCameraPreset
      tags: [cameras]
      summary: Save PTZ preset
      parameters:
        - name: cameraId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PtzPresetInput"
      responses:
        "201":
          description: Preset saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PtzPreset"
...
  /stream-deck/profiles:
    get:
      operationId: getStreamDeckProfiles
      tags: [devices]
      summary: List Stream Deck profiles
      responses:
        "200":
          description: Profiles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StreamDeckProfile"
    post:
      operationId: createStreamDeckProfile
      tags: [devices]
      summary: Create a Stream Deck profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StreamDeckProfileInput"
      responses:
        "201":
          description: Profile created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StreamDeckProfile"
...
components:
  schemas:
...
    Camera:
      type: object
      required: [cameraId, name, manufacturer, adapter, ipAddress, rtspUrl, connectionState, createdAt]
      properties:
        cameraId:
          type: string
        name:
          type: string
        manufacturer:
          type: string
          enum: [axis, canon, panasonic, sony, ptzoptics, birddog, aver, bolin, onvif, generic-rtsp]
        adapter:
          type: string
        model:
          type: ["string", "null"]
        ipAddress:
          type: string
        rtspUrl:
          type: string
        username:
          type: ["string", "null"]
        streamProfile:
          type: ["string", "null"]
        codec:
          type: ["string", "null"]
        resolution:
          type: ["string", "null"]
        frameRate:
          type: ["integer", "null"]
        connectionState:
          type: string
          enum: [connected, disconnected, connecting, error]
        healthState:
          type: ["string", "null"]
        firmwareVersion:
          type: ["string", "null"]
        serialNumber:
          type: ["string", "null"]
        bitRate:
          type: ["integer", "null"]
        droppedFrames:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: ["string", "null"]
          format: date-time
    CameraInput:
      type: object
      required: [name, manufacturer, adapter, ipAddress, rtspUrl]
      properties:
        name:
          type: string
        manufacturer:
          type: string
          enum: [axis, canon, panasonic, sony, ptzoptics, birddog, aver, bolin, onvif, generic-rtsp]
        adapter:
          type: string
        model:
          type: string
        ipAddress:
          type: string
        rtspUrl:
          type: string
        username:
          type: string
        password:
          type: string
        streamProfile:
          type: string
        codec:
          type: string
        resolution:
          type: string
        frameRate:
          type: integer
    CameraConnectionResult:
      type: object
      required: [success, connectionState]
      properties:
        success:
          type: boolean
        connectionState:
          type: string
        message:
          type: ["string", "null"]
        cameraId:
          type: string
    CameraTestResult:
      type: object
      required: [success, latencyMs]
      properties:
        success:
          type: boolean
        latencyMs:
          type: ["integer", "null"]
        resolution:
          type: ["string", "null"]
        frameRate:
          type: ["number", "null"]
        bitRate:
          type: ["integer", "null"]
        message:
          type: ["string", "null"]
    PtzCommand:
      type: object
      required: [action]
      properties:
        action:
          type: string
          enum: [pan-left, pan-right, tilt-up, tilt-down, zoom-in, zoom-out, stop, home, preset-recall, preset-save, emergency-wide]
        panSpeed:
          type: ["number", "null"]
        tiltSpeed:
          type: ["number", "null"]
        zoomSpeed:
          type: ["number", "null"]
        presetId:
          type: ["string", "null"]
        presetName:
          type: ["string", "null"]
    PtzPreset:
      type: object
      required: [presetId, name, cameraId]
      properties:
        presetId:
          type: string
        name:
          type: string
        cameraId:
          type: string
        description:
          type: ["string", "null"]
        isDefault:
          type: boolean
    PtzPresetInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
        isDefault:
          type: boolean
...
    StreamDeckProfile:
      type: object
      required: [profileId, name, sport, buttons, createdAt]
      properties:
        profileId:
          type: string
        name:
          type: string
        sport:
          type: string
          enum: [gaelic-football, hurling, ladies-football, camogie, custom]
        isDefault:
          type: boolean
        buttons:
          type: array
          items:
            $ref: "#/components/schemas/StreamDeckButton"
        createdAt:
          type: string
          format: date-time
    StreamDeckProfileInput:
      type: object
      required: [name, sport, buttons]
      properties:
        name:
          type: string
        sport:
          type: string
          enum: [gaelic-football, hurling, ladies-football, camogie, custom]
        isDefault:
          type: boolean
        buttons:
          type: array
          items:
            $ref: "#/components/schemas/StreamDeckButton"
    StreamDeckButton:
      type: object
      required: [buttonNumber, label, eventType]
      properties:
        buttonNumber:
          type: integer
        label:
          type: string
        icon:
          type: ["string", "null"]
        colour:
          type: ["string", "null"]
        eventType:
          type: string
        team:
          type: ["string", "null"]
          enum: [home, away, null]
        scoreEffect:
          type: ["string", "null"]
        overlayEffect:
          type: ["string", "null"]
        clipRequest:
          type: boolean
        enabled:
          type: boolean
...
    OperationResult:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        message:
          type: ["string", "null"]
        error:
          type: ["string", "null"]